+++ examples/quickstart1.py
@@ -1,0 +1,26 @@
+import numpy as np
+import matthewplotlib as mp
+
+xs = np.linspace(-2*np.pi, +2*np.pi, 150)
+
+plot = mp.axes(
+    mp.scatter(
+        (xs, 1.0 * np.cos(xs), "red"),
+        (xs, 0.9 * np.cos(xs - 0.33 * np.pi), "magenta"),
+        (xs, 0.8 * np.cos(xs - 0.66 * np.pi), "blue"),
+        (xs, 0.7 * np.cos(xs - 1.00 * np.pi), "cyan"),
+        (xs, 0.8 * np.cos(xs - 1.33 * np.pi), "green"),
+        (xs, 0.9 * np.cos(xs - 1.66 * np.pi), "yellow"),
+        width=75,
+        height=10,
+        yrange=(-1,1),
+    ),
+    title=" y = cos(x + 2πk/6) ",
+    xlabel="x",
+    ylabel="y",
+)
+
+print(plot)
+
+plot.saveimg("images/quickstart.png")
+